pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "chesscoach"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "chess>=1.10",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "requests>=2.33.1",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
]

[project.scripts]
chess-coach = "chesscoach.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8",
]

[tool.pytest.ini_options]
markers = [
    "integration: requires stockfish binary installed on the system",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.10",
]
